requirements.txt
aiofiles==24.1.0
aiogram==2.25.1
aiogram-media-group==0.5.1
aiogram-timepicker==0.2.1
aiogram_calendar==0.1.1
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiomysql==0.2.0
aioredis==2.0.1
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
APScheduler==3.11.0
apscheduler-di==0.1.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
easydict==1.13
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
magic-filter==1.0.12
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.4
onnx==1.17.0
onnxruntime==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.2.1
prettytable==3.16.0
propcache==0.3.1
protobuf==6.30.2
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
rodi==2.0.8
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.4
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
wcwidth==0.2.13
yarl==1.20.0
